remove PeerStore registration logic, move TrackerFuncs
